automatically backed up by learn

+++ db/migrate/003_create_haunted_houses.rb
@@ -5,8 +5,8 @@ class CreateCostumeStores < ActiveRecord::Migration[4.2]
     create_table :haunted_houses do |t|
       t.string :name
       t.string :location
-      t.string :theme 
-      t.integer :price 
+      t.string :theme
+      t.float :price
       t.boolean :still_in_business
       t.datetime :opening_time
       t.datetime :closing_time
